revert back to int64_t base data type

+++ benchmark.cpp
@@ -24,14 +24,14 @@ int main(int argc, char** argv)
 #define MAX_PROBLEM_SIZE 1 << 28  //  256M
    std::vector<int64_t> problem_sizes{ MAX_PROBLEM_SIZE >> 5, MAX_PROBLEM_SIZE >> 4, MAX_PROBLEM_SIZE >> 3, MAX_PROBLEM_SIZE >> 2, MAX_PROBLEM_SIZE >> 1, MAX_PROBLEM_SIZE};
    
-   float *A = (float *)malloc(sizeof(float) * MAX_PROBLEM_SIZE);
+   int64_t *A = (int64_t *)malloc(sizeof(int64_t) * MAX_PROBLEM_SIZE);
 
    int n_problems = problem_sizes.size();
 
    /* For each test size */
    for (int64_t n : problem_sizes) 
    {
-      float t;
+      int64_t t;
       printf("Working on problem size N=%lld \n", n);
 
       // invoke user code to set up the problem
