+++ sum_vector.cpp
@@ -9,13 +9,13 @@
 #include "sums.h"
 
 void 
-setup(int64_t N, float A[])
+setup(int64_t N, int64_t A[])
 {
    printf(" inside sum_vector problem_setup, N=%lld \n", N);
 }
 
-float
-sum(int64_t N, float A[])
+int64_t
+sum(int64_t N, int64_t A[])
 {
    printf(" inside sum_vector perform_sum, N=%lld \n", N);
