+++ sums.h
@@ -1,7 +1,7 @@
 #ifndef __sum_h_
 #define __sum_h_
 
-extern void setup(int64_t N, float A[]);
-extern float sum(int64_t N, float A[]);
+extern void setup(int64_t N, int64_t A[]);
+extern int64_t sum(int64_t N, int64_t A[]);
 
 #endif
